rename the component and its implementation
